Convert indicators to package
