+++ technical/indicators/cycle_indicators.py
@@ -1,0 +1,14 @@
+"""
+Cycle indicators
+"""
+
+########################################
+#
+# Cycle Indicator Functions
+#
+
+# HT_DCPERIOD          Hilbert Transform - Dominant Cycle Period
+# HT_DCPHASE           Hilbert Transform - Dominant Cycle Phase
+# HT_PHASOR            Hilbert Transform - Phasor Components
+# HT_SINE              Hilbert Transform - SineWave
+# HT_TRENDMODE         Hilbert Transform - Trend vs Cycle Mode
