+++ technical/indicators/indicators.py
@@ -7,228 +7,6 @@ from numpy.core.records import ndarray
 from pandas import Series, DataFrame
 from math import log
 
-
-########################################
-#
-# Overlap Studies Functions
-#
-
-# BBANDS               Bollinger Bands
-def bollinger_bands(dataframe, period=21, stdv=2, field='close', colum_prefix="bb") -> DataFrame:
-    from pyti.bollinger_bands import lower_bollinger_band, middle_bollinger_band, upper_bollinger_band
-    dataframe["{}_lower".format(colum_prefix)] = lower_bollinger_band(dataframe[field], period, stdv)
-    dataframe["{}_middle".format(colum_prefix)] = middle_bollinger_band(dataframe[field], period, stdv)
-    dataframe["{}_upper".format(colum_prefix)] = upper_bollinger_band(dataframe[field], period, stdv)
-
-    return dataframe
-
-
-# DEMA                 Double Exponential Moving Average
-
-# EMA                  Exponential Moving Average
-def ema(dataframe, period, field='close'):
-    import talib.abstract as ta
-    return ta.EMA(dataframe, timeperiod=period, price=field)
-
-
-# HT_TRENDLINE         Hilbert Transform - Instantaneous Trendline
-# KAMA                 Kaufman Adaptive Moving Average
-# MA                   Moving average
-# MAMA                 MESA Adaptive Moving Average
-# MAVP                 Moving average with variable period
-# MIDPOINT             MidPoint over period
-# MIDPRICE             Midpoint Price over period
-# SAR                  Parabolic SAR
-# SAREXT               Parabolic SAR - Extended
-
-# SMA                  Simple Moving Average
-def sma(dataframe, period, field='close'):
-    import talib.abstract as ta
-    return ta.SMA(dataframe, timeperiod=period, price=field)
-
-
-# T3                   Triple Exponential Moving Average (T3)
-
-# TEMA                 Triple Exponential Moving Average
-def tema(dataframe, period, field='close'):
-    import talib.abstract as ta
-    return ta.TEMA(dataframe, timeperiod=period, price=field)
-
-
-# TRIMA                Triangular Moving Average
-# WMA                  Weighted Moving Average
-
-# Other Overlap Studies Functions
-def hull_moving_average(dataframe, period, field='close') -> ndarray:
-    from pyti.hull_moving_average import hull_moving_average as hma
-    return hma(dataframe[field], period)
-
-
-def vwma(df, window):
-    return df.apply(lambda x: x.close * x.volume, axis=1).rolling(window).sum() / df.volume.rolling(window).sum()
-
-
-def zema(dataframe, period, field='close'):
-    """
-    zero lag ema
-    :param dataframe:
-    :param period:
-    :param field:
-    :return:
-    """
-    dataframe = dataframe.copy()
-    dataframe['ema1'] = ema(dataframe, period, field)
-    dataframe['ema2'] = ema(dataframe, period, 'ema1')
-    dataframe['d'] = dataframe['ema1'] - dataframe['ema2']
-    dataframe['zema'] = dataframe['ema1'] + dataframe['d']
-    return dataframe['zema']
-
-
-########################################
-#
-# Momentum Indicator Functions
-#
-
-# ADX                  Average Directional Movement Index
-# ADXR                 Average Directional Movement Index Rating
-# APO                  Absolute Price Oscillator
-
-# AROON                Aroon
-def aroon(dataframe, period=25, field='close', colum_prefix="aroon") -> DataFrame:
-    from pyti.aroon import aroon_up as up
-    from pyti.aroon import aroon_down as down
-    dataframe["{}_up".format(colum_prefix)] = up(dataframe[field], period)
-    dataframe["{}_down".format(colum_prefix)] = down(dataframe[field], period)
-    return dataframe
-
-
-# AROONOSC             Aroon Oscillator
-# BOP                  Balance Of Power
-
-# CCI                  Commodity Channel Index
-def cci(dataframe, period) -> ndarray:
-    from pyti.commodity_channel_index import commodity_channel_index
-
-    return commodity_channel_index(dataframe['close'], dataframe['high'], dataframe['low'], period)
-
-
-# CMO                  Chande Momentum Oscillator
-def cmo(dataframe, period, field='close') -> ndarray:
-    from pyti.chande_momentum_oscillator import chande_momentum_oscillator
-    return chande_momentum_oscillator(dataframe[field], period)
-
-
-# DX                   Directional Movement Index
-# MACD                 Moving Average Convergence/Divergence
-# MACDEXT              MACD with controllable MA type
-# MACDFIX              Moving Average Convergence/Divergence Fix 12/26
-# MFI                  Money Flow Index
-# MINUS_DI             Minus Directional Indicator
-# MINUS_DM             Minus Directional Movement
-
-# MOM                  Momentum
-def momentum(dataframe, field='close', period=9):
-    from pyti.momentum import momentum as m
-    return m(dataframe[field], period)
-
-
-# PLUS_DI              Plus Directional Indicator
-# PLUS_DM              Plus Directional Movement
-# PPO                  Percentage Price Oscillator
-# ROC                  Rate of change : ((price/prevPrice)-1)*100
-# ROCP                 Rate of change Percentage: (price-prevPrice)/prevPrice
-# ROCR                 Rate of change ratio: (price/prevPrice)
-# ROCR100              Rate of change ratio 100 scale: (price/prevPrice)*100
-# RSI                  Relative Strength Index
-# STOCH                Stochastic
-# STOCHF               Stochastic Fast
-# STOCHRSI             Stochastic Relative Strength Index
-# TRIX                 1-day Rate-Of-Change (ROC) of a Triple Smooth EMA
-
-# ULTOSC               Ultimate Oscillator
-def ultimate_oscilator(dataframe):
-    from pyti.ultimate_oscillator import ultimate_oscillator as uo
-    uo(dataframe['close'], dataframe['low'])
-
-
-# WILLR                Williams' %R
-def williams_percent(dataframe, period=14, field='close'):
-    highest_high = dataframe[field].rolling(period).max()
-    lowest_low = dataframe[field].rolling(period).min()
-    wr = (highest_high - dataframe[field]) / (highest_high - lowest_low) * -100
-    return wr
-
-
-########################################
-#
-# Volume Indicator Functions
-#
-
-# AD                   Chaikin A/D Line
-def accumulation_distribution(dataframe) -> ndarray:
-    from pyti.accumulation_distribution import accumulation_distribution as acd
-
-    return acd(dataframe['close'], dataframe['high'], dataframe['low'], dataframe['volume'])
-
-
-# ADOSC                Chaikin A/D Oscillator
-# OBV                  On Balance Volume
-
-# Other Volume Indicator Functions
-def chaikin_money_flow(dataframe, period=21):
-    mfm = (
-        (dataframe['close'] - dataframe['low']) - (dataframe['high'] - dataframe['close'])
-        / (dataframe['high'] - dataframe['low'])
-    )
-    mfv = mfm * dataframe['volume']
-    cmf = mfv.rolling(period).sum() / dataframe['volume'].rolling(period).sum()
-    return cmf
-
-cmf = chaikin_money_flow
-
-########################################
-#
-# Cycle Indicator Functions
-#
-
-# HT_DCPERIOD          Hilbert Transform - Dominant Cycle Period
-# HT_DCPHASE           Hilbert Transform - Dominant Cycle Phase
-# HT_PHASOR            Hilbert Transform - Phasor Components
-# HT_SINE              Hilbert Transform - SineWave
-# HT_TRENDMODE         Hilbert Transform - Trend vs Cycle Mode
-
-
-########################################
-#
-# Price Transform Functions
-#
-
-# AVGPRICE             Average Price
-# MEDPRICE             Median Price
-# TYPPRICE             Typical Price
-# WCLPRICE             Weighted Close Price
-
-
-########################################
-#
-# Volatility Indicator Functions
-#
-
-# ATR                  Average True Range
-def atr(dataframe, period, field='close') -> ndarray:
-    from pyti.average_true_range import average_true_range
-    return average_true_range(dataframe[field], period)
-
-
-def atr_percent(dataframe, period, field='close') -> ndarray:
-    from pyti.average_true_range_percent import average_true_range_percent
-    return average_true_range_percent(dataframe[field], period)
-
-
-# NATR                 Normalized Average True Range
-# TRANGE               True Range
-
-
 ########################################
 #
 # Pattern Recognition Functions
