+++ technical/indicators/price_transform.py
@@ -1,0 +1,13 @@
+"""
+Cycle indicators
+"""
+
+########################################
+#
+# Price Transform Functions
+#
+
+# AVGPRICE             Average Price
+# MEDPRICE             Median Price
+# TYPPRICE             Typical Price
+# WCLPRICE             Weighted Close Price
